added functionality behind the camera. The take picture feature now works and the video functionality partly works.
Added the functionality behind the text formatted feature. When the text is selected the user can select if they want to used the Bold, Italic, Underline, or strikethrough functions.

Added the text size dropbox feature. This feature does not completely work. When the text formatting feature us clicked the text size with revert to 12.